Move all source files out of the build directory tree
Also a new directory was created (resources) to contain all fonts, music and other
stuff and a tuxpak command was added to create a main.pak.

The game.py of the Python demo programs are also moved into the src tree. During
the build a copy of game.py is placed in the respective directories.

Please note that a copy of main.pak must be copied to the directory where the
binary of the program resides. This is a temporary solution.