Add a logging / diagnostics facility.
With this release we can add logging / diagnostic in the TuxCap code. It is
enabled by using the --log-level command line option. Here is a command
line example to run hungarr:
    build/bin/Hungarr --log-level=sexyappbase,board=2 --log=-
The formal log level syntax is <FACIL> [ '=' <LEVEL> ] ( ',' <FACIL> [ '=' <LEVEL> ] )*
<FACIL> is just a name, <LEVEL> is a number (can even be 0x prefixed hex).

In other words you can enable one or more "facilities". The level says
how much detail you want to see. The higher the level the more detail.
So that means that the TuxCap programmer gets to decide what is "more
detail". Mostly it also relates to how often a log entry can be given.

The --log option argument is the filename of the log file. Use - to send
the logging to stdout (as in the example above).

The names of the facilities are determined by the programmer of TuxCap. And
as a suggestion it is the name of a class (in lowercase) or the name of
a feature (such as keyboard).