Redoing the viewport and window creation. Fixes the mysteriously clipped game images.
Quite a bit of code was modified (also in the previous commits) to figure
out how surfaces and images play a role. It turned out that the clipping
was done based on the real window, not on the game dimension.
In the new setup there is a "game" surface with the game dimension,
and that one is used by DDImage and it has the correct width and height
of the game.